Fix: game loop time; Fix: var type

+++ src/main/java/org/example/keenmarksmanfx/IBulletFactory.java
@@ -1,5 +1,5 @@
 package org.example.keenmarksmanfx;
 
 interface IBulletFactory {
-    void createBullet(float x, float y);
+    void createBullet(double x, double y);
 }
